+++ packages/patternfly-4/react-table/src/components/Table/utils/decorators/textCenter.ts
@@ -1,0 +1,1 @@
+export const textCenter = () => ({ textCenter: true });