+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.6...@patternfly/react-docs@4.10.7) (2019-08-15)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.10.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.5...@patternfly/react-docs@4.10.6) (2019-08-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.10.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.10.4...@patternfly/react-docs@4.10.5) (2019-08-13)
 
 **Note:** Version bump only for package @patternfly/react-docs